Copy all config files instead of just ${MIX_ENV}
Without `runtime.exs` the port binding to `0.0.0.0` doesn't work,
causing the docker container instance of the server to be unreachable
from outside.

Credentials are optional right now for a Job, don't expect one when
running a Job.

Defaulted a Run's log to `[]`, for some reason I was getting errors when
rendering a log should afaik have a log list.