Fix selection of different jobs in workflow view
The phx-ignore on the editor meant we didn't see server-side content changes (such as those triggered by changing selected job) so the editor was forever out of sync. I've made the editor a bit less clever (usually for the best) so that instead of managing its own hidden child, it is given the id of the bound hidden input. This seems to work well.