Tighten up the mutation observer
Two issues here. Firstly, there's a mutation being triggered on every keystroke (even if the attribute value hasn't actually changed! This must be LiveView re-rendering even with phx-update set to ignore.

Secondly, when selecting different jobs, the editor will only update the adpator happens to be different. I'm now explicitly tracking the job id in the editor DOm so that I know when it's changed (thanks Stu for the inspiration!)