Fix Builder tab issue after switching jobs
There is something weird happening with LiveView and Alpine, when
updated components are copied they are supposed to be cloned by Alpine.

But in the cases where they _shouldn't_ change (as in their tree is
identical) it appears that the `to` element doesn't exist or isn't
rendered.

My guess is that the component either is cloned correctly or there is
something else going on with updating data on the parent elements.

I do find it strange how the `panel_content` apepars to be fine.