Fix issue involving new jobs not getting a default cron expression
When creating a new Job, and selecting "Cron" as the trigger - and then
not setting any of the frequency/time options the job either didn't save
(backend error) or saved with a `null` `cron_expression` in the database.

This addresses all of the above, by default a `cron_expression` is
'Daily'. In addition a regression was found when changing an existing
triggers cron expression.