Add SCHEMAS_PATH for use in production
In production/release environments, the `pwd` changes to `/app/bin`
which impacts the default (releative) `./priv/..` lookups.

In the case with schemas, we expect them to be in `/app/priv/schemas`
and so we allow it to be set via an ENV variable - and default it in the
Docker image.