Ensure that JSON Schema inputs are rendered in the correct order
Due to Elixirs maps not having a gaurenteed order (and therefore
ExJsonSchema) a schema written to be in a certain order isn't
necessarily going to be rendered out in the form in the same order.

By leveraging `Jason.OrderedObject` access behaviour we get the order of
the `properties` keys when setting the list of `fields` on
`%Lightning.Credentials.Schema{}`.

Fixes #685