Remove conflicting show_modal and hide_model functions
We have a few modal implementations, however they now conflict with the
addition of the CoreComponents functions.

I commented the `LightningWeb.Components.Modal` module out for now, this
should be the new single modal across the application.

I also renamed the `show|hide` functions in `ModalPortal` to
`on_show|hide` as not to conflict with CoreComponents.