Prevent pushHistoryPatches from using an old location
There are situations where the browsers location.href hasn't been
changed yet but the component has already mounted, leaving the diagram
to navigate with a selection query param on the workflow list url
instead of the workflow edit url.

So we store the url in the liveview process and set it as a `data-` attr
on the hook. Since we can't trust the browsers location we have to
track it elsewhere.