Improvements to selecting nodes
While this was improved not so long ago, I found there were still
moments that caused the baseUrl used to decide if the browser should
navigate to a new url was out of date - the result was clicking on a
node would sometimes have no effect.

This approach listens to all `phx:navigate` requests to update the
`baseUrl` which appears to be much more reliable.