Preload and keep the attempts workorder and project
When manually running a job, the output pane would flat
(white-out/rerender).

This was caused by the URLs not having the required preloads
(`...workflow.project` in this case).

What was happening is that we _were_ preloading correctly, however when
the attempt is updated a new version of it (without preloads) replaces
it and the LiveView crashes.

It settles down when the Attempt is complete and it no longer updated.

Fixes #1550