Stop cancel/confirm placeholder job from propagating click events (#1564)
When clicking cancel or confirm, the click event was being propagated up
and calling `handleSelection`, resulting in two `pushHistoryPatch`
events being fired within a few usecs.

LiveView fails to render some (if not all) diffs calculated from the workflow state
store. In subsequent renders an error would be thrown when those
elements need to be changed and aren't on the DOM.

This _may_ be a bug in LiveView, or intended behaviour. But for now, we
shouldn't be pushing a new URL twice anyway. I've added a
`stopPropagation` to both the commit and cancel buttons on the
PlaceholderJob component.

Fixes #1537