Fix Docs Panel (#2198)
* run prettier on docs panel

* docs-panel: tweak margins to take up maximum width

in vertical alignment, various paddings mean that the component doe not take all the available wwidth, and so the scrollbars are in the wrong place

* docs-panel: stop the tabs creating a scrollbar if there isnt enoguh space

this created a not very useful double scrollbar in small screens. Now if there isn't enough space, some of the tab is inaccessible. This just has to be how it is - there simply isn't enough screen space

* docs-panel: restore overflow-hidden to the editors parent

* Update CHANGELOG


---------

Co-authored-by: Stuart Corbishley <corbish@gmail.com>