Added workflow concurrency queries
There are two new queries `in_progress_window` and `eligible_for_claim`.
Both can be found in the `Lightning.Runs.Query` module.

The effect of these queries in practice is that once a workflow has it's
`concurrency` field set, once that number of runs is claimed or started
has been reached (even though they arrived first) will not be considered
for the next run to be claimed.

The `FifoRunQueue` (the default) queue, now uses the
`eligible_for_claim` query to determine which runs to claim next.