Set max_frame_size to 11MB
`max_frame_size` needs to be set at the socket in endpoint, not in the
Endpoint config.

Setting it to 11MB statically for now, no easy way to make this dynamic
since `socket` doesn't appear to use plug builder (i.e. can't use
Replug).

What happens when a worker sends a payload that is too big it drops the
connection on the floor, not exactly what we want but ok for now I
guess. The worker will see a timeout error on the socket.

Fixes #2339