Ignore Plug.Conn.InvalidQueryError
This error occurs when someone sends a bad UTF-8 encoded query param.
There is nothing that we can do but send back a 400 error, which is what
happens in both dev and production environments

So there is no need to register this error in Sentry either, as these
kinds of errors are usually associated with bots.