Upgrade Elixir to 1.18 and OTP to 27 (#2748)
* Upgrade Elixir to 1.18

* Add build_local script for Docker build management

This script facilitates local Docker builds with options for cache management and build progress display. It ensures the existence of a buildx builder and configures it for the build process.

* Use NodeSource to install NodeJS

* Bump Dockerfile elixir versions

* Add support for custom Erlang VM flags in Docker build

Updated the Dockerfile to include an argument for ERL_FLAGS, allowing users to pass custom Erlang VM flags during the build process. Enhanced the build_local script to accept a --platform option and display usage information, improving flexibility for local Docker builds.

- Added "coveralls.json" task to mix.exs for test coverage reporting.


* Refactor CircleCI configuration and add ci_tests script

- Updated CircleCI config to replace coveralls.json execution with a new ci_tests script for improved test coverage reporting.
- Introduced ci_tests script to run Elixir tests, handle retries for failed tests, and generate JUnit XML reports for CI environments.

---------

Co-authored-by: Rory McKinley <rorymckinley@gmail.com>
Co-authored-by: Rogerio Pontual <rppontual@gmail.com>