Forking simpler non-LiveView macro controlled React hook component
The original implementation is now called HeexReactComponent, which the
jsx macro now uses. The existing parts of the application use this right
now.

Making room for the new CollaborativeEditor component with needing
LiveView to be involved at a rendering level. So that means we won't
have the application/json props etc. The 'new' ReactComponent hook uses
simple attr plucking and a replaceEqualDeep to give us the same prop
passing behaviour.